requirements/requirements-api.txt
flask==2.1.2
Werkzeug==2.2.2
tensorflow==2.13.1
numpy==1.22.3
pillow==9.1.0
waitress==3.0.0
